'修改ajax及某些操作排除get请求'

+++ comment/views.py
@@ -1,5 +1,9 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.auth.decorators import login_required, permission_required
+from django.views.decorators.csrf import csrf_protect, ensure_csrf_cookie, csrf_exempt
+from django.core.exceptions import SuspiciousOperation
+from django.http import HttpResponse
+
 
 # Create your views here.
 
@@ -12,8 +16,8 @@ from blog.models import Post
 @login_required(login_url='/account/login/')
 @permission_required('comment.add_comment')
 def post_comment(request, post_pk, parent_comment_pk=0):
-    post = get_object_or_404(Post, pk=post_pk)
     if request.method == 'POST':
+        post = get_object_or_404(Post, pk=post_pk)
         form = CommentForm(request.POST)
         if form.is_valid():
             if parent_comment_pk != '0' and parent_comment_pk != 0:
@@ -35,33 +39,40 @@ def post_comment(request, post_pk, parent_comment_pk=0):
                     }
 
             return render(request, '_comments.html', context=context)
-    # return redirect(post)
+        return HttpResponse('ajax_error')
+    raise SuspiciousOperation
+
 
 def comment_delete(request, pk):
-    comment = get_object_or_404(Comment, pk=pk)
-    post = comment.post
-    comment.delete()
-    form = CommentForm()
-    comment_list = post.get_comments()
-    context = {
-        'post': post,
-        'form': form,
-        'comment_list': comment_list
-    }
-    # return redirect(post)
-    return render(request, '_comments.html', context=context)
+    if request.method == 'POST':
+        comment = get_object_or_404(Comment, pk=pk)
+        post = comment.post
+        comment.delete()
+        form = CommentForm()
+        comment_list = post.get_comments()
+        context = {
+            'post': post,
+            'form': form,
+            'comment_list': comment_list
+        }
+        # return redirect(post)
+        return render(request, '_comments.html', context=context)
+    raise SuspiciousOperation
+
 
 def comment_enable(request, pk):
-    comment = get_object_or_404(Comment, pk=pk)
-    post = comment.post
-    comment.enable = not comment.enable
-    comment.save()
-    form = CommentForm()
-    comment_list = post.get_comments()
-    context = {
-        'post': post,
-        'form': form,
-        'comment_list': comment_list
-    }
-    # return redirect(comment.post)
-    return render(request, '_comments.html', context=context)
+    if request.method == 'POST':
+        comment = get_object_or_404(Comment, pk=pk)
+        post = comment.post
+        comment.enable = not comment.enable
+        comment.save()
+        form = CommentForm()
+        comment_list = post.get_comments()
+        context = {
+            'post': post,
+            'form': form,
+            'comment_list': comment_list
+        }
+        # return redirect(comment.post)
+        return render(request, '_comments.html', context=context)
+    raise SuspiciousOperation
